tests: clarified how to load cryptonote_utils
