--- tests/emjs/tmp-cnutil.node.js
@@ -1,1 +1,0 @@
-const mymonero = require("../../");
