+++ tests/index.node.js
@@ -1,0 +1,19 @@
+"use strict";
+const mymonero = require("../");
+// const assert = require("assert");
+
+var public_key =
+	"904e49462268d771cc1649084c35aa1296bfb214880fe2e7f373620a3e2ba597";
+var private_key =
+	"52aa4c69b93b780885c9d7f51e6fd5795904962c61a2e07437e130784846f70d";
+
+var nettype = mymonero.nettype_utils.network_type.MAINNET;
+
+mymonero.monero_utils.OnceModuleReady(function(Module)
+{
+	var decoded = mymonero.monero_utils.decode_address(
+		"49qwWM9y7j1fvaBK684Y5sMbN8MZ3XwDLcSaqcKwjh5W9kn9qFigPBNBwzdq6TCAm2gKxQWrdZuEZQBMjQodi9cNRHuCbTr",
+		nettype,
+	);
+	console.log("decoded", decoded)
+})
